1288 - Move TenantThreadPoolTaskExecutor under tenant-api
