+++ server/libs/atlas/atlas-sync-executor/build.gradle.kts
@@ -9,4 +9,5 @@ dependencies {
     implementation(project(":server:libs:atlas:atlas-worker:atlas-worker-impl"))
     implementation(project(":server:libs:core:commons:commons-util"))
     implementation(project(":server:libs:core:message:message-broker:message-broker-sync"))
+    implementation(project(":server:libs:core:tenant:tenant-api"))
 }
